Refactor Partner loading code
