+++ components/home/partners/index.tsx
@@ -2,7 +2,7 @@ import * as S from "./styles";
 import LogoList from "components/logo-list";
 import csstrings from "content/strings.json";
 import enstrings from "content/strings-en.json";
-import { PortalPartner } from "lib/portal-types";
+import { PortalPartner } from "lib/airtable/partner";
 
 interface Props {
   partners: readonly PortalPartner[];
