+++ lib/site-data.ts
@@ -1,15 +1,14 @@
 import { getAllVideos, PortalVideo } from "./cedu";
 import { Article, getArticleIndex } from "./related-blog-posts";
 import { Field, getAllSkills } from "./airtable/skills";
-import * as Airtable from "./airtable-import";
 import * as Local from "./local-data";
 import apm from "elastic-apm-node";
 import { enableAPMLogging } from "./apm";
 import { getAllUsers, PortalUser } from "./airtable/user";
-import { PortalPartner } from "./portal-types";
 import { getAllProjects, PortalProject } from "./airtable/project";
 import { getAllEvents, PortalEvent } from "./airtable/event";
 import { getAllOpportunities, PortalOpportunity } from "./airtable/opportunity";
+import { getAllPartners, PortalPartner } from "./airtable/partner";
 
 enableAPMLogging();
 
@@ -42,7 +41,7 @@ const ProductionDataSource: DataSource = {
   opportunities: getAllOpportunities,
   users: getAllUsers,
   events: getAllEvents,
-  partners: Airtable.getAllPartners,
+  partners: getAllPartners,
   videos: getAllVideos,
   blogPosts: getArticleIndex,
   skills: getAllSkills,
@@ -61,7 +60,8 @@ const SampleDataSource: DataSource = {
 
 async function loadSiteData(): Promise<SiteData> {
   const forceLocal = !!process.env.DATA_SOURCE_LOCAL;
-  const useLocalData = forceLocal || !Airtable.isAvailable;
+  const haveAirtable = !!process.env.AIRTABLE_API_KEY;
+  const useLocalData = forceLocal || !haveAirtable;
   const isProductionBuild = process.env.VERCEL_ENV === "production";
 
   if (useLocalData && isProductionBuild) {
